another size_ vs int type mismatch found
git-svn-id: https://code.boutroslab.oicr.on.ca/svn/BoutrosLab@73431 aa8168d8-c5e5-4185-88a0-f3e152ed764a

+++ src/cget_trinucleotide_counts.cpp
@@ -14,7 +14,7 @@ Rcpp::NumericVector cget_trinucleotide_counts(std::vector<std::string> key, std:
     std::map<std::string, long int> count;
 
     //initialize
-    for (int i = 0; i < key.size(); i++) 
+    for (size_t i = 0; i < key.size(); i++) 
         count.insert( std::pair<std::string, long int>(key[i], 0) );
 
     std::ifstream f(chr.c_str());
@@ -55,7 +55,7 @@ Rcpp::NumericVector cget_trinucleotide_counts(std::vector<std::string> key, std:
 
     Rcpp::NumericVector count_vec(key.size());
 
-    for (int i = 0; i < key.size(); ++i)
+    for (size_t i = 0; i < key.size(); ++i)
         count_vec[i] = count.at(key[i]);
 
     delete[] tri_nucl_c;
